logmind: Phase 2 (W3 final): bin/clud-bug.js to 3-line shim importing dist/cli/index.js
